requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
nest_asyncio
playwright
google-generativeai
pandas
pdfplumber
requests
beautifulsoup4
pydantic